Use `moonwalk` for `dua a`

+++ src/interactive/app/tests/utils.rs
@@ -40,7 +40,7 @@ pub fn node_by_name(app: &TerminalApp, name: impl AsRef<OsStr>) -> &EntryData {
 pub fn index_by_name_and_size(
     app: &TerminalApp,
     name: impl AsRef<OsStr>,
-    size: Option<u128>,
+    size: Option<u64>,
 ) -> TreeIndex {
     let name = name.as_ref();
     let t: Vec<_> = app
@@ -289,7 +289,7 @@ pub fn sample_02_tree() -> Tree {
     tree
 }
 
-pub fn make_add_node(t: &mut Tree) -> impl FnMut(&str, u128, Option<NodeIndex>) -> NodeIndex + '_ {
+pub fn make_add_node(t: &mut Tree) -> impl FnMut(&str, u64, Option<NodeIndex>) -> NodeIndex + '_ {
     move |name, size, maybe_from_idx| {
         let n = t.add_node(EntryData {
             name: PathBuf::from(name),
